feat: Move some file to common folder
